test(architectures): 拆分 RegistryInitializationHookBaseTests 辅助类型
- 拆分 RegistryInitializationHookBaseTests 末尾的测试辅助类型到同目录独立文件以消除 MA0048

- 更新 TestRegistry 的公开集合暴露方式为只读接口以规避 MA0016

- 补充新测试辅助类型的 XML 文档并保持原有测试行为不变
