+++ GFramework.Core.Tests/Architectures/RegistryInitializationHookBaseTests.cs
@@ -1,11 +1,5 @@
-using System.Reflection;
-using GFramework.Core.Abstractions.Architectures;
+using System;
 using GFramework.Core.Abstractions.Enums;
-using GFramework.Core.Abstractions.Lifecycle;
-using GFramework.Core.Abstractions.Model;
-using GFramework.Core.Abstractions.Systems;
-using GFramework.Core.Abstractions.Utility;
-using GFramework.Core.Architectures;
 
 namespace GFramework.Core.Tests.Architectures;
 
@@ -114,293 +108,5 @@ public class RegistryInitializationHookBaseTests
         hook.OnPhase(ArchitecturePhase.AfterSystemInit, architecture);
 
         Assert.That(registry.RegisteredConfigs.Count, Is.EqualTo(2));
-    }
-}
-
-/// <summary>
-/// 测试用的注册表初始化钩子实现
-/// </summary>
-public class TestRegistryInitializationHook : RegistryInitializationHookBase<TestRegistry, string>
-{
-    public TestRegistryInitializationHook(
-        IEnumerable<string> configs,
-        ArchitecturePhase targetPhase = ArchitecturePhase.AfterSystemInit)
-        : base(configs, targetPhase)
-    {
-    }
-
-    protected override void RegisterConfig(TestRegistry registry, string config)
-    {
-        registry.Register(config);
-    }
-}
-
-/// <summary>
-/// 测试用的注册表类
-/// </summary>
-public class TestRegistry : IUtility
-{
-    public List<string> RegisteredConfigs { get; } = new();
-
-    public void Register(string config)
-    {
-        RegisteredConfigs.Add(config);
-    }
-}
-
-/// <summary>
-/// 测试用的架构类（包含注册表）
-/// </summary>
-public class TestArchitectureWithRegistry : IArchitecture
-{
-    private readonly TestRegistry _registry;
-
-    public TestArchitectureWithRegistry(TestRegistry registry)
-    {
-        _registry = registry;
-        Context = new TestArchitectureContextWithRegistry(registry);
-    }
-
-    public Action<IServiceCollection>? Configurator { get; }
-
-    public IArchitectureContext Context { get; }
-    Action<IServiceCollection>? IArchitecture.Configurator => Configurator;
-
-    T IArchitecture.RegisterSystem<T>(T system)
-    {
-        throw new NotSupportedException();
-    }
-
-    T IArchitecture.RegisterModel<T>(T model)
-    {
-        throw new NotSupportedException();
-    }
-
-    T IArchitecture.RegisterUtility<T>(T utility)
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterCqrsPipelineBehavior<TBehavior>() where TBehavior : class
-    {
-        throw new NotSupportedException();
-    }
-
-    /// <summary>
-    ///     测试替身未实现显式程序集 CQRS 处理器接入入口。
-    /// </summary>
-    /// <param name="assembly">包含 CQRS 处理器或生成注册器的程序集。</param>
-    /// <exception cref="NotSupportedException">该测试替身不参与 CQRS 程序集接入路径验证。</exception>
-    public void RegisterCqrsHandlersFromAssembly(Assembly assembly)
-    {
-        throw new NotSupportedException();
-    }
-
-    /// <summary>
-    ///     测试替身未实现显式程序集 CQRS 处理器接入入口。
-    /// </summary>
-    /// <param name="assemblies">要接入的程序集集合。</param>
-    /// <exception cref="NotSupportedException">该测试替身不参与 CQRS 程序集接入路径验证。</exception>
-    public void RegisterCqrsHandlersFromAssemblies(IEnumerable<Assembly> assemblies)
-    {
-        throw new NotSupportedException();
-    }
-
-    public IArchitectureModule InstallModule(IArchitectureModule module)
-    {
-        throw new NotSupportedException();
-    }
-
-    IArchitectureLifecycleHook IArchitecture.RegisterLifecycleHook(IArchitectureLifecycleHook hook)
-    {
-        throw new NotSupportedException();
-    }
-
-    Task IArchitecture.WaitUntilReadyAsync()
-    {
-        return WaitUntilReadyAsync();
-    }
-
-    public void RegisterUtility<T>(Action<T>? onCreated = default(Action<T>?)) where T : class, IUtility
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterModel<T>(Action<T>? onCreated = default(Action<T>?)) where T : class, IModel
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterSystem<T>(Action<T>? onCreated = default(Action<T>?)) where T : class, ISystem
-    {
-        throw new NotSupportedException();
-    }
-
-    public void Initialize()
-    {
-    }
-
-    public void Destroy()
-    {
-        throw new NotSupportedException();
-    }
-
-    Task IAsyncInitializable.InitializeAsync()
-    {
-        return InitializeAsync();
-    }
-
-    ValueTask IAsyncDestroyable.DestroyAsync()
-    {
-        return DestroyAsync();
-    }
-
-    public Task WaitUntilReadyAsync()
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterLifecycleHook(IArchitectureLifecycleHook hook)
-    {
-    }
-
-    public Task InitializeAsync()
-    {
-        throw new NotSupportedException();
-    }
-
-    public ValueTask DestroyAsync()
-    {
-        throw new NotSupportedException();
-    }
-}
-
-/// <summary>
-/// 测试用的架构上下文类（包含注册表）
-/// </summary>
-public class TestArchitectureContextWithRegistry : TestArchitectureContext
-{
-    private readonly TestRegistry _registry;
-
-    public TestArchitectureContextWithRegistry(TestRegistry registry)
-    {
-        _registry = registry;
-    }
-
-    public override TUtility? GetUtility<TUtility>() where TUtility : class
-    {
-        if (typeof(TUtility) == typeof(TestRegistry))
-        {
-            return _registry as TUtility;
-        }
-
-        return base.GetUtility<TUtility>();
-    }
-}
-
-/// <summary>
-/// 测试用的架构类（不包含注册表）
-/// </summary>
-public class TestArchitectureWithoutRegistry : IArchitecture
-{
-    public TestArchitectureWithoutRegistry()
-    {
-        Context = new TestArchitectureContext();
-    }
-
-    public IArchitectureContext Context { get; }
-    public Action<IServiceCollection>? Configurator { get; }
-
-    T IArchitecture.RegisterSystem<T>(T system)
-    {
-        throw new NotSupportedException();
-    }
-
-    T IArchitecture.RegisterModel<T>(T model)
-    {
-        throw new NotSupportedException();
-    }
-
-    T IArchitecture.RegisterUtility<T>(T utility)
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterCqrsPipelineBehavior<TBehavior>() where TBehavior : class
-    {
-        throw new NotSupportedException();
-    }
-
-    /// <summary>
-    ///     测试替身未实现显式程序集 CQRS 处理器接入入口。
-    /// </summary>
-    /// <param name="assembly">包含 CQRS 处理器或生成注册器的程序集。</param>
-    /// <exception cref="NotSupportedException">该测试替身不参与 CQRS 程序集接入路径验证。</exception>
-    public void RegisterCqrsHandlersFromAssembly(Assembly assembly)
-    {
-        throw new NotSupportedException();
-    }
-
-    /// <summary>
-    ///     测试替身未实现显式程序集 CQRS 处理器接入入口。
-    /// </summary>
-    /// <param name="assemblies">要接入的程序集集合。</param>
-    /// <exception cref="NotSupportedException">该测试替身不参与 CQRS 程序集接入路径验证。</exception>
-    public void RegisterCqrsHandlersFromAssemblies(IEnumerable<Assembly> assemblies)
-    {
-        throw new NotSupportedException();
-    }
-
-    public IArchitectureModule InstallModule(IArchitectureModule module)
-    {
-        throw new NotSupportedException();
-    }
-
-    IArchitectureLifecycleHook IArchitecture.RegisterLifecycleHook(IArchitectureLifecycleHook hook)
-    {
-        throw new NotSupportedException();
-    }
-
-    public Task WaitUntilReadyAsync()
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterUtility<T>(Action<T>? onCreated = default(Action<T>?)) where T : class, IUtility
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterModel<T>(Action<T>? onCreated = default(Action<T>?)) where T : class, IModel
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterSystem<T>(Action<T>? onCreated = default(Action<T>?)) where T : class, ISystem
-    {
-        throw new NotSupportedException();
-    }
-
-    public void Initialize()
-    {
-    }
-
-    public Task InitializeAsync()
-    {
-        throw new NotSupportedException();
-    }
-
-    public ValueTask DestroyAsync()
-    {
-        throw new NotSupportedException();
-    }
-
-    public void Destroy()
-    {
-        throw new NotSupportedException();
-    }
-
-    public void RegisterLifecycleHook(IArchitectureLifecycleHook hook)
-    {
     }
 }
